pyproject.toml
[project]
name = "python-async-pdf-chat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "langchain>=1.2.4",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.7",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
    "redis>=7.1.0",
    "rq>=2.6.1",
    "rq-dashboard>=0.8.6",
    "uvicorn>=0.40.0",
]
